Parall

See merge request Young/compiler2024-nel!86

+++ src/mir/Instruction.java
@@ -402,7 +402,7 @@ public class Instruction extends User {
             }
         }
 
-        public void setProbability(float probability) {
+        public void setProbability(double probability) {
             this.probability = probability;
         }
